update setup script

+++ setup.sh
@@ -3,7 +3,9 @@
 mkdir -p $HOME/.local/bin
 
 if [[ -f $HOME/.bashrc ]]; then
-  mv $HOME/.bashrc $HOME/.bashrc.bak
+  if [[ ! -L $HOME/.bashrc ]]; then
+    mv $HOME/.bashrc $HOME/.bashrc.bak
+  fi
 fi
 
 stow .
@@ -13,3 +15,9 @@ source $HOME/.bashrc
 if [[ ! -x "$(command -v mise)" ]]; then
   uget mise
 fi
+
+source $HOME/.local/installers/build-deps
+
+mise i -y
+
+source $HOME/.local/installers/nerd-fonts Hack
